pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cod-tempo-real"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dynamic = ["version"]
dependencies = [
    "numpy>=2.3.2",
    "plotly>=6.2.0",
    "polars>=1.31.0",
    "pyarrow>=21.0.0",
    "pyjson5>=2.0.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "statsmodels>=0.14.5",
]

[tool.hatch.version]
path = "app/__init__.py"

[tool.hatch.build.targets.wheel]
include = [
  "app/",
  "main.py"
]

[tool.ruff]
line-length = 80

[project.scripts]
rest2-ons = "main:main"